[worker] Document response stream disconnected error (#18567)
* [worker] Document response stream disconnected error

This change documents the `responseStreamDisconnected` error `outcome`
seen in CloudFlare worker logs and error metrics.

Previously, this outcome doesn't appear to be documented, but presents
as an error in metrics and presents as a non-error entry in the
CloudFlare Worker Logs. This change aims to reduce confusion going forward.

Changes introducing `responseStreamDisconnected`: https://github.com/cloudflare/workerd/pull/2503

* Update src/content/docs/workers/observability/errors.mdx

* Moved content under metrics heading

---------

Co-authored-by: ToriLindsay <tgalatro@cloudflare.com>